requirements.txt
requests~=2.32.3
python-dotenv~=1.0.1
py-cord~=2.5.0